requirements.txt
certifi==2020.12.5
chardet==4.0.0
Faker==6.6.2
idna==2.10
python-dateutil==2.8.1
python-logstash==0.4.6
requests==2.25.1
six==1.15.0
text-unidecode==1.3
urllib3==1.26.4
